# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StorySmith Review Console API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects": {
      "get": {
        "summary": "List Projects",
        "operationId": "list_projects_projects_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSummary"
                  },
                  "title": "Response List Projects Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}": {
      "get": {
        "summary": "Get Project",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/approve": {
      "post": {
        "summary": "Approve Project",
        "operationId": "approve_project_projects__project_id__approve_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/reject": {
      "post": {
        "summary": "Reject Project",
        "operationId": "reject_project_projects__project_id__reject_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs": {
      "post": {
        "summary": "Trigger Run",
        "operationId": "trigger_run_runs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shows": {
      "post": {
        "summary": "Create Show",
        "description": "User-authored, frozen cast (Amendment 02) -- no LLM involved. The\nuser's own words become CharacterRef.description/personality directly;\nthis only generates and locks in one reference portrait per character.",
        "operationId": "create_show_shows_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShowRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Shows",
        "operationId": "list_shows_shows_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShowSummary"
                  },
                  "title": "Response List Shows Shows Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assets/view": {
      "get": {
        "summary": "View Asset",
        "description": "Generic authenticated asset proxy (Amendment 02) -- streams any\nstorage URI's bytes through the server via the same StoragePort.get()\nevery agent already uses, sidestepping local storage's non-browser-\nfetchable local:// scheme entirely. Used client-side (fetch + Blob +\nobject URL, since neither <img src> nor a plain link can carry the\nbearer header) for the show-creation avatar gallery and the final-video\ndownload button.",
        "operationId": "view_asset_assets_view_get",
        "parameters": [
          {
            "name": "uri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uri"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateShowRequest": {
        "properties": {
          "show_id": {
            "type": "string",
            "title": "Show Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "art_style": {
            "type": "string",
            "title": "Art Style"
          },
          "palette": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Palette",
            "default": []
          },
          "mood": {
            "type": "string",
            "title": "Mood",
            "default": "cheerful"
          },
          "tempo_bpm": {
            "type": "integer",
            "title": "Tempo Bpm",
            "default": 100
          },
          "aspect_ratio": {
            "type": "string",
            "title": "Aspect Ratio",
            "default": "9:16"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution",
            "default": "1080x1920"
          },
          "pacing_rules": {
            "type": "string",
            "title": "Pacing Rules",
            "default": ""
          },
          "characters": {
            "items": {
              "$ref": "#/components/schemas/CharacterInput"
            },
            "type": "array",
            "title": "Characters"
          }
        },
        "type": "object",
        "required": [
          "show_id",
          "name",
          "art_style",
          "characters"
        ],
        "title": "CreateShowRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Mode": {
        "type": "string",
        "enum": [
          "rhyme",
          "topical"
        ],
        "title": "Mode"
      },
      "ProjectDetail": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "brief": {
            "type": "string",
            "title": "Brief"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd"
          },
          "qa_reports": {
            "items": {
              "$ref": "#/components/schemas/QAReportOut"
            },
            "type": "array",
            "title": "Qa Reports"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetOut"
            },
            "type": "array",
            "title": "Assets"
          },
          "published_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Url"
          },
          "show_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Id"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "mode",
          "brief",
          "status",
          "title",
          "total_cost_usd",
          "qa_reports",
          "assets",
          "published_url",
          "show_id"
        ],
        "title": "ProjectDetail"
      },
      "ProjectSummary": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "brief": {
            "type": "string",
            "title": "Brief"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "show_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Id"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "status",
          "mode",
          "brief",
          "title",
          "total_cost_usd",
          "updated_at",
          "show_id"
        ],
        "title": "ProjectSummary"
      },
...
      "RunRequest": {
        "properties": {
          "brief": {
            "type": "string",
            "title": "Brief"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode",
            "default": "rhyme"
          },
          "show_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Id"
          }
        },
        "type": "object",
        "required": [
          "brief"
        ],
        "title": "RunRequest"
      },
      "RunResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "RunResponse"
      },
      "ShowDetail": {
        "properties": {
          "show_id": {
            "type": "string",
            "title": "Show Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "art_style": {
            "type": "string",
            "title": "Art Style"
          },
          "characters": {
            "items": {
              "$ref": "#/components/schemas/CharacterOut"
            },
            "type": "array",
            "title": "Characters"
          }
        },
        "type": "object",
        "required": [
          "show_id",
          "name",
          "art_style",
          "characters"
        ],
        "title": "ShowDetail"
      },
      "ShowSummary": {
        "properties": {
          "show_id": {
            "type": "string",
            "title": "Show Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "show_id",
          "name",
          "created_at"
        ],
        "title": "ShowSummary"
      },
...
    }
  }
}